Split HIR property checks (#1305)
Move property signature validation, static verdict detection, and property diagnostic sample/help text out of the HIR monolith into hir/properties.rs.

Update compiler source monolith reporting and ratchet ceilings for the new compiler.hir property module.

Refs #1254.
