+++ scripts/ci/report-compiler-source-monoliths.py
@@ -41,6 +41,7 @@ BOUNDARY_MAP: dict[str, list[str]] = {
     "model.rs": ["compiler.hir"],
     "new_project.rs": ["compiler.commands"],
     "ownership.rs": ["compiler.hir"],
+    "properties.rs": ["compiler.hir"],
     "project.rs": ["compiler.package_graph", "compiler.commands", "compiler.evidence"],
     "registry.rs": ["compiler.package_graph"],
     "signatures.rs": ["compiler.hir"],
